Extract `Resource` as a separate reusable structure
