+++ tests/registries/test_decorators.py
@@ -3,7 +3,8 @@ import pytest
 import kopf
 from kopf.reactor.causation import CREATE, UPDATE, DELETE
 from kopf.reactor.handling import subregistry_var
-from kopf.reactor.registries import Resource, SimpleRegistry, GlobalRegistry
+from kopf.reactor.registries import SimpleRegistry, GlobalRegistry
+from kopf.structs.resources import Resource
 
 
 def test_on_create_minimal(mocker):
