+++ tests/registries/test_requires_finalizer.py
@@ -1,7 +1,8 @@
 import pytest
 
 import kopf
-from kopf.reactor.registries import Resource, GlobalRegistry
+from kopf.reactor.registries import GlobalRegistry
+from kopf.structs.resources import Resource
 
 
 OBJECT_BODY = {
